refactor(sysfs): split sysfs.c into per-domain modules
Break the 1266-line monolithic sysfs.c into focused units while moving
manual GC control out of the production attribute surface.

  sysfs.c           1266 -> 265   core: version/region_info/perm_info/daxheap_bufid + group/init
  sysfs_me.{c,h}    new          ME inspection + per-CPU stats (me_info, poll_stats, fine_stats, ...)
  sysfs_gc.{c,h}    new          GC monitoring (deleg_info, gc_status)
  sysfs_nrht.{c,h}  new          NRHT chain-depth histogram
  sysfs_internal.h  new          shared sbi_list/lock + get_sbi/find_by_node helpers
  sysfs_debug.{c,h} extended     gc_trigger/stop/pause/restart relocated into debug subgroup
                                 alongside existing fault injection (me_freeze_heartbeat,
                                 me_sync_is_holder)

Helper naming in sysfs_me.c unified to verb-noun form:
  me_state_str             -> me_state_name
  me_tag_for               -> me_format_tag
  me_info_emit_one         -> me_emit_instance
  me_stats_aggregate       -> me_aggregate_stats
  me_fine_stats_emit_buckets -> me_emit_buckets

Tests updated to /sys/fs/marufs/debug/gc_* paths:
  test_local_multinode.sh  (13 spots)
  test_chown_race.c, test_dupname.c, test_overlap.c, test_gc_deleg.c

Also includes pre-staged ME crash-detection scaffolding consumed by the
debug subgroup: me.h / me_order.c / me_request.c expose debug_freeze_poll
hooks; tests/test_me_crash.sh exercises freeze + sync recovery end-to-end.
