+++ marufs_kernel/src/sysfs_nrht.h
@@ -1,0 +1,13 @@
+/* SPDX-License-Identifier: GPL-2.0-only */
+/*
+ * sysfs_nrht.h - NRHT (Non-Resident Hash Table) sysfs attributes.
+ */
+
+#ifndef _MARUFS_SYSFS_NRHT_H
+#define _MARUFS_SYSFS_NRHT_H
+
+#include <linux/sysfs.h>
+
+extern struct kobj_attribute nrht_chain_depth_attr;
+
+#endif /* _MARUFS_SYSFS_NRHT_H */
